Update README.md
This commit adds a README file that outlines the purpose and structure of the HackerRank_Java repository. It highlights the beginner-friendly nature of the solutions, their adherence to HackerRank’s instructions, and provides guidance on how to use and contribute to the project. The README also includes licensing details and acknowledges HackerRank as the source of the problem sets.